requirements.txt
streamlit
google-genai
google-cloud-vision
pillow
python-dotenv
